Update triage.yml
It is recommended to explicitly specify the event type for the workflow. In this case, you can use issues as the event type. But you need to use issue instead of issues in the event configuration. The issues keyword should be replaced with issue since you want the workflow to trigger when an issue is opened or reopened.